Multiple QOL improvements to the renderer and its architecture
- Fixed two critical static member bugs where IGraphicsPipeline::m_Specification and IShader::m_Name were declared inline static, causing all instances to share a single value. Both are now instance members. IGraphicsPipeline::Create now passes the spec through to the constructor. COpenGLShader now assigns m_Name from the filepath stem, fixing GetName() returning empty and shader library lookups storing shaders under an empty key. Also fixed scalar delete on an array allocation in CRenderer2D::Shutdown; m_QuadVertexBufferBase is now correctly freed with delete[].

- Consolidated the duplicated four-vertex write block across all DrawQuad and DrawRotatedQuad overloads into a single private EmitQuadVertices helper. The batch overflow check now lives exclusively there. Rotated quad overloads now delegate to their DrawQuad counterparts rather than duplicating texture slot lookup. Removed the dead SetFloat("u_TilingFactor") call from Flush and the redundant RenderPacket.Texture assignment, as texture binding already happens in the preceding m_TextureSlots loop.

- Removed the EditorViewportCamera BeginFrame overload and its include from CRenderer2D. The call site in CScene::OnRenderEditor now extracts the view-projection matrix before calling in.

- Removed SetX methods from IShader entirely. Sampler initialization is now handled automatically in COpenGLShader via reflection after program linking — any sampler uniforms found are initialized with sequential slot indices, requiring no per-shader setup. The Renderer2DQuad shader now declares the texture array with explicit register(t0) and vk::binding(1, 0) annotations for future Vulkan compatibility.

- Introduced IRHIFactory and FRHI, consolidating all RHI object creation from per-type static Create methods with repeated switch statements into a single factory interface. FRHI holds the factory instance, exposes it via FRHI::GetFactory(), and contains the sole backend-selection switch in FRHI::Initialize(). The static Create methods and their backend includes have been removed from all RHICore types, eliminating the dependency from the abstract RHI layer back onto its own implementations.

+++ Engine/Source/Runtime/OpenGLRHI/OpenGLCommandList.cpp
@@ -3,14 +3,13 @@
 #include "OpenGLGraphicsPipeline.h"
 #include "OpenGLIndexBuffer.h"
 #include "OpenGLTexture.h"
+#include "OpenGLUniformBuffer.h"
 #include "OpenGLVertexBuffer.h"
 #include "Debug/Profiler.h"
 #include "Renderer/Renderer.h"
 
 #include <glad/glad.h>
 #include <tracy/TracyOpenGL.hpp>
-
-#include "OpenGLUniformBuffer.h"
 
 void COpenGLCommandList::SetGraphicsPipeline(const std::shared_ptr<IGraphicsPipeline>& GraphicsPipeline)
 {
@@ -65,7 +64,7 @@ void COpenGLCommandList::DrawIndexed(uint32 IndexCount, uint32 InstanceCount, ui
     verifyEnginef(m_BoundVertexBuffer, "{} requires a bound vertex buffer!", __FUNCTION__)
     verifyEnginef(m_BoundIndexBuffer, "{} requires a bound index buffer!", __FUNCTION__)
 
-    const int32 Count = IndexCount == 0 ? m_BoundIndexBuffer->GetCount() : IndexCount;
+    const int32 Count = IndexCount == 0 ? static_cast<int32>(m_BoundIndexBuffer->GetCount()) : IndexCount;
     
     glDrawElements(GL_TRIANGLES, Count, GL_UNSIGNED_INT, nullptr);
 }
